+++ package.json
@@ -46,6 +46,7 @@
     ]
   },
   "devDependencies": {
+    "eslint-config-react": "^1.1.7",
     "eslint-plugin-react": "^7.37.2",
     "react-router-dom": "^6.28.0",
     "tailwindcss": "^3.4.14"
