+++ src/App.js
@@ -1,8 +1,8 @@
 import "./App.css";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
 import PageLayout from "./navigation/PageLayout";
-import Classes from "./classes/Classes";
-import Class from "./classes/class/Class";
+import Classes from "./pages/classes/Classes";
+import ClassDetail from "./pages/class-detail/ClassDetail";
 import Error404 from "./navigation/Error404";
 
 function App() {
@@ -11,7 +11,7 @@ function App() {
       <Routes>
         <Route element={<PageLayout/>}>
           <Route path="/classes" element={<Classes/>}/>
-          <Route path="/classes/:classURL" element={<Class/>}/>
+          <Route path="/classes/:classURL" element={<ClassDetail/>}/>
           <Route path="*" element={<Error404/>}/>
         </Route>
       </Routes>
