--- src/components/H1.js
@@ -1,5 +1,0 @@
-function H1({ children }) {
-  return <h1 className="text-5xl text-neutral-black font-extrabold font-display">{children}</h1>;
-}
-
-export default H1;
