--- src/components/H2.js
@@ -1,5 +1,0 @@
-function H2({ children }) {
-  return <h2 className="text-3xl text-neutral-black font-bold font-display">{children}</h2>;
-}
-
-export default H2;
