+++ src/pages/classes/example.js
